scrape query sql tops out at our assumed max recurseDepth. Also...
break out query debugger into separate file ignored by coverage
