+++ src/tools/store/querier-entrypoint.ts
@@ -1,70 +1,15 @@
-import * as readlineSync from 'readline-sync'
 import { existsSync } from '../../util/fs'
 import { UninitializedDatabaseError } from '../../util/errors'
 import { Database } from './database'
 import { selectOrderedScrapers } from './queries'
 import { typecheckQueryArguments } from '../../util/typechecking.runtime'
+import { queryExecutionDebugger } from './query-debugger'
 // type imports
+import { DebuggerView } from './query-debugger'
 import { Settings } from '../../settings'
 import { ScraperName } from '../../settings/config/types'
-import {
-  SelectedRow as OrderedScrapersRow,
-  SelectedRowWithDebugInfo as OrderedScrapersRowWithDebug
-} from './queries/select-ordered-scrapers'
+import { SelectedRow as OrderedScrapersRow } from './queries/select-ordered-scrapers'
 import { ArgumentTypes } from '../../util/types'
-
-type DebuggerView = (keyof OrderedScrapersRowWithDebug)[]
-const queryExecutionDebugger = (
-  scrapers: string[],
-  settings: Settings,
-  result: OrderedScrapersRow[],
-  debuggerView: DebuggerView
-) => {
-  // we are cheating with the type here because function overloading in typescript doesnt play nicely with higher order functions
-  const resultAsDebug = (result as any) as OrderedScrapersRowWithDebug[]
-  readlineSync.question(
-    'This is an internal debugging method. It will hault execution until there is input. Press [enter] to continue.',
-    { mask: '' }
-  )
-
-  const scraperConfigs = scrapers.map(s => settings.flatConfig.getOrThrow(s))
-  const lowestDepth = Math.max(...scraperConfigs.map(s => s.depth))
-
-  const resultAsDepthMap: Map<number, OrderedScrapersRow[]> = resultAsDebug.reduce((acc, row) => {
-    acc
-      .set(row.recurseDepth, acc.get(row.recurseDepth) || [])
-      .get(row.recurseDepth)!
-      .push(row)
-    return acc
-  }, new Map())
-  const resultSplitByDepth = Array.from(resultAsDepthMap.entries())
-    // the need for this filter might be because of an error with this shite
-    .filter(entry => entry[0] <= lowestDepth)
-    .sort((a, b) => a[0] - b[0])
-    .map(entry => entry[1])
-
-  let key = ''
-  let index = 0
-  while (key !== 'q') {
-    const resultAtDepth = resultSplitByDepth[index]
-    const viewable = resultAtDepth.map(row =>
-      // TODO move this to util/object.ts (filterObjectKeys)?
-      debuggerView.reduce((acc: { [column: string]: any }, column) => {
-        acc[column] = (row as any)[column]
-        return acc
-      }, {})
-    )
-    console.clear() // eslint-disable-line no-console
-    console.table(viewable) // eslint-disable-line no-console
-    key = readlineSync.keyIn(
-      'Press [s] to move forward, press [w] to move backward, press [q] to exit the debugger',
-      { limit: 'swq' }
-    )
-
-    if (key === 's' && index < resultSplitByDepth.length - 1) index++
-    else if (key === 'w' && index > 0) index--
-  }
-}
 
 export const EXECUTION_DEBUGGER_VIEW = Symbol.for('query-execution-stepper')
 export type QueryArguments = {
@@ -108,15 +53,13 @@ const querierFactory = (settings: Settings): QueryFn => {
 
     const scrapersInQuery = scrapers.concat(groupBy || [])
 
-    const preparedStatment = selectOrderedScrapers(flatConfig, database)(
-      Array.from(new Set(scrapersInQuery)),
-      Boolean(debuggerView)
-    )
+    const sqlCompiler = selectOrderedScrapers(flatConfig, database)
+    const preparedStmt = sqlCompiler(Array.from(new Set(scrapersInQuery)), false)
 
     return () => {
-      const result = preparedStatment()
+      if (debuggerView) queryExecutionDebugger(scrapersInQuery, sqlCompiler, debuggerView)
 
-      if (debuggerView) queryExecutionDebugger(scrapersInQuery, settings, result, debuggerView)
+      const result = preparedStmt()
 
       const groupedResults = []
       let groupCount = 0
