+++ webpack.config.js
@@ -36,7 +36,7 @@ module.exports = (env, { mode = 'development' } = {}) => ({
       {
         test: content => env === 'coverage' && /\.ts$/.test(content),
         include: path.resolve(__dirname, 'src'),
-        exclude: /node_modules|\.test.ts$/,
+        exclude: [/node_modules/, /\.test\.ts$/, /query-debugger\.ts$/],
         loader: 'istanbul-instrumenter-loader',
         options: { esModules: true },
         enforce: 'post'
